fix for no "ammo" being removed
Small changes to fix "ammo" not being removed when using "Respawn Loadout", it first saves a default loadout/price when spawn, and then when respawn with "Respawn Loadout" it changed the backup loadout to the default one so if you dont have enough money it changes to that. And also sets the "old price" to default price.